Fencing_Clock redundancy efficiency changes
Redid how the time string is built in Fencing_Clock to the fastest way I could find, switched how the Point_Displays build their strings to match, and added in a redundancy check for Fencing_Clock to not bother with updates outside of if a whole number of seconds has changed. Bought about 400 microseconds in the average, timer-off, lights-off case!